backend/requirements.txt
# backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
pyodbc>=5.0.0
pandas>=2.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
chromadb>=0.5.0
sentence-transformers>=2.7.0
